refactor(filesystem): move filesystem classes into a new top-level package
